refactor: Split palettes / field-i18n / formatters out of ui-helpers
