+++ app/charts/shared/axis-width-band.tsx
@@ -5,7 +5,7 @@ import { useEffect, useRef } from "react";
 import { ColumnsState } from "@/charts/column/columns-state";
 import { useChartState } from "@/charts/shared/use-chart-state";
 import { useChartTheme } from "@/charts/shared/use-chart-theme";
-import { useTimeFormatUnit } from "@/configurator/components/ui-helpers";
+import { useTimeFormatUnit } from "@/formatters";
 
 export const AxisWidthBand = () => {
   const ref = useRef<SVGGElement>(null);
