+++ app/charts/shared/axis-width-linear.tsx
@@ -6,7 +6,7 @@ import { BarsState } from "@/charts/bar/bars-state";
 import { ScatterplotState } from "@/charts/scatterplot/scatterplot-state";
 import { useChartState } from "@/charts/shared/use-chart-state";
 import { useChartTheme } from "@/charts/shared/use-chart-theme";
-import { useFormatNumber } from "@/configurator/components/ui-helpers";
+import { useFormatNumber } from "@/formatters";
 import { estimateTextWidth } from "@/utils/estimate-text-width";
 import { MaybeTooltip } from "@/utils/maybe-tooltip";
